filter exchanges based on subzones
