+++ web/geo/generate-exchanges-to-exclude.js
@@ -1,35 +1,40 @@
 const { getJSON, writeJSON, fileExists } = require('./utilities');
 const exchangeConfig = require('../../config/exchanges.json');
 
-const generateExchangesToIgnore = (OUT_PATH, fc) => {
+const generateExchangesToIgnore = (OUT_PATH, zonesConfig) => {
   console.log(`Generating new excluded-aggregated-exchanges.json...`); // eslint-disable-line no-console
-  const fcCombined = {
-    ...fc,
-    features: fc.features.filter((feature) => {
-      try {
-        return feature.properties.isCombined;
-      } catch (e) {
-        console.log('Error: ', e, 'Feature: ', feature); // eslint-disable-line no-console
-      }
-    }),
-  };
 
-  const countryKeysToExclude = fcCombined.features.map((feature) => feature.properties.countryKey);
-
-  const unCombinedExchanges = Object.keys(exchangeConfig).filter((key) => {
-    const split = key.split('->');
-    const zoneOne = split[0].slice(0, 2);
-    const zoneTwo = split[1].slice(0, 2);
-    if (key.length > 6 && (countryKeysToExclude.includes(zoneOne) || countryKeysToExclude.includes(zoneTwo))) {
+  const countryKeysToExclude = Object.keys(zonesConfig).filter((key) => {
+    if (zonesConfig[key].subZoneNames?.length > 0) {
       return key;
     }
   });
 
+  //Create a list of the exchange keys that we don't want to display in a country view
+  const unCombinedExchanges = Object.keys(exchangeConfig).filter((key) => {
+    const split = key.split('->');
+    const zoneOne = split[0];
+    const zoneTwo = split[1];
+
+    const subzoneSplitOne = zoneOne.split('-');
+    const subzoneSplitTwo = zoneTwo.split('-');
+    if (
+      (zoneOne.includes('-') && countryKeysToExclude.includes(subzoneSplitOne[0])) ||
+      (zoneTwo.includes('-') && countryKeysToExclude.includes(subzoneSplitTwo[0]))
+    ) {
+      return key;
+    }
+  });
+
+  //Create a list of the exchange keys that we don't want to display in the zone view
   const countryExchangesWithSubzones = Object.keys(exchangeConfig).filter((key) => {
     const split = key.split('->');
-    const zoneOne = split[0].slice(0, 2);
-    const zoneTwo = split[1].slice(0, 2);
-    if (key.length === 6 && (countryKeysToExclude.includes(zoneOne) || countryKeysToExclude.includes(zoneTwo))) {
+    const zoneOne = split[0];
+    const zoneTwo = split[1];
+    if (
+      (!zoneOne.includes('-') && countryKeysToExclude.includes(zoneOne)) ||
+      (!zoneTwo.includes('-') && countryKeysToExclude.includes(zoneTwo))
+    ) {
       return key;
     }
   });
